refactor(frontend): improve FormattedNumberInput behavior and validation
- Change input type from number to text to support thousands separators and localized input formatting.
- Support configurable decimal precision using the numberOfDecimals prop instead of a hardcoded value of 2.
- Implement handleBlur and handleFocus to manage active edit state, parse localized raw input, and re-format on blur.
- Fix min/max validation checks by parsing numeric values from the formatted string representation before checking constraints.
