--- src/components/inputs/FormattedNumberInput.js
@@ -1,124 +1,0 @@
-import React, { Component } from "react";
-import TextInput from "./TextInput";
-import { injectIntl } from "react-intl";
-import { formatMessage, formatMessageWithValues } from "../../helpers/i18n";
-import { withModulesManager } from "@openimis/fe-core";
-
-class FormattedNumberInput extends Component {
-  constructor(props) {
-    super(props);
-
-    this.pricesAreDecimal = props.modulesManager.getConf("fe-core", "pricesAreDecimal", true);
-
-    this.state = {
-      isEdited: false,
-      rawValue: props.value != null
-        ? this.formatNumber(props.value, props.intl)
-        : "",
-    };
-  }
-
-  componentDidUpdate(prevProps) {
-    // Si la valeur du serveur change et que l’utilisateur n’édite pas, on reformate
-    if (prevProps.value !== this.props.value && !this.state.isEdited) {
-      this.setState({
-        rawValue: this.props.value != null
-          ? this.formatNumber(this.props.value, this.props.intl)
-          : "",
-      });
-    }
-  }
-
-  formatNumber = (value, intl) => {
-    if (value == null || isNaN(value)) return "";
-    return new Intl.NumberFormat(intl, {
-      minimumFractionDigits: this.pricesAreDecimal ? 2 : 0,
-      maximumFractionDigits: this.pricesAreDecimal ? 2 : 0,
-    }).format(value);
-  };
-
-  handleKeyPress = (event) => {
-    const { allowDecimals = true } = this.props;
-    if (event.key === "." && !allowDecimals) {
-      event.preventDefault();
-    }
-  };  
-
-  handleChange = (val) => {
-    const raw = val;
-    this.setState({ rawValue: raw });
-
-    const normalized = raw.replace(/\s/g, "").replace(",", ".");
-    const value = parseFloat(normalized);
-    this.props.onChange(isNaN(value) ? undefined : value);
-  };
-
-  handleBlur = () => {
-    const { intl, displayNa } = this.props;
-    const { rawValue } = this.state;
-
-    this.setState({ isEdited: false });
-
-    if ((rawValue === "" || isNaN(Number(rawValue))) && displayNa) {
-      this.setState({
-        rawValue: formatMessage(intl, this.props.module, "core.NumberInput.notApplicable"),
-      });
-      return;
-    }
-
-    const number = parseFloat(rawValue.replace(/\s/g, "").replace(",", "."));
-    if (isNaN(number)) {
-      this.setState({ rawValue: "" });
-      return;
-    }
-
-    this.setState({ rawValue: this.formatNumber(number, intl) });
-  };
-
-  handleFocus = () => {
-    this.setState({ isEdited: true });
-  };
-
-  render() {
-    const {
-      intl,
-      module = "core",
-      min = null,
-      max = null,
-      error,
-      allowDecimals = true,
-      ...others
-    } = this.props;
-
-    let inputProps = {
-      ...this.props.inputProps,
-      type: "text",
-      onKeyPress: this.handleKeyPress,
-    };
-
-    let err = error;
-
-    const numericValue = parseFloat(this.state.rawValue.replace(/\s/g, "").replace(",", "."));
-    if (min != null && numericValue < min) {
-      err = formatMessageWithValues(intl, module, "validation.minValue", { value: numericValue, min });
-    }
-    if (max != null && numericValue > max) {
-      err = formatMessageWithValues(intl, module, "validation.maxValue", { value: numericValue, max });
-    }
-
-    return (
-      <TextInput
-        {...others}
-        module={module}
-        value={this.state.rawValue}
-        error={err}
-        inputProps={inputProps}
-        onChange={this.handleChange}
-        onBlur={this.handleBlur}
-        onFocus={this.handleFocus}
-      />
-    );
-  }
-}
-
-export default withModulesManager(injectIntl(FormattedNumberInput));
